+++ Sandboxie/apps/common/RunBrowser.cpp
@@ -134,7 +134,7 @@ void CRunBrowser::OnNo()
 
 CString CRunBrowser::GetTopicUrl(const CString &topic)
 {
-    return L"https://www.sandboxie.com/index.php?" + topic;
+    return L"https://xanasoft.com/Sandboxie/" + topic;
 }
 
 
@@ -155,5 +155,5 @@ void CRunBrowser::OpenHelp(CWnd *pParentWnd, const CString &topic)
 
 void CRunBrowser::OpenForum(CWnd *pParentWnd)
 {
-    CRunBrowser x(pParentWnd, L"http://forums.sandboxie.com/phpBB3/");
+    CRunBrowser x(pParentWnd, L"https://forum.xanasoft.com/");
 }