+++ Sandboxie/common/my_version.h
@@ -20,8 +20,8 @@
 #ifndef _MY_VERSION_H
 #define _MY_VERSION_H
 
-#define MY_VERSION_BINARY       5,42,0
-#define MY_VERSION_STRING       "5.42.0"
+#define MY_VERSION_BINARY       5,42,1
+#define MY_VERSION_STRING       "5.42.1"
 #define MY_VERSION_COMPAT		"5.42"
 
 // These #defines are used by either Resource Compiler, or by NSIC installer