+++ SandboxiePlus/QSbieAPI/Sandboxie/SbieIni.h
@@ -5,12 +5,14 @@
 
 #include "../SbieError.h"
 
-class QSBIEAPI_EXPORT CIniSection: public QObject
+class QSBIEAPI_EXPORT CSbieIni: public QObject
 {
 	Q_OBJECT
 public:
-	CIniSection(const QString& Section, class CSbieAPI* pAPI, QObject* parent = 0);
-	virtual ~CIniSection();
+	CSbieIni(const QString& Section, class CSbieAPI* pAPI, QObject* parent = 0);
+	virtual ~CSbieIni();
+
+	virtual QString		GetName() const { return m_Name; }
 
 	virtual SB_STATUS SetText(const QString& Setting, const QString& Value);
 	virtual SB_STATUS SetNum(const QString& Setting, int Value);
@@ -22,15 +24,22 @@ public:
 	virtual __int64 GetNum64(const QString& Setting, __int64 Default = 0) const;
 	virtual bool GetBool(const QString& Setting, bool Default = false) const;
 
-	virtual QStringList GetTextList(const QString &Setting, bool withBrackets = false);
+	virtual QStringList GetTextList(const QString &Setting, bool withTemplates = true) const;
+	virtual SB_STATUS UpdateTextList(const QString &Setting, const QStringList& List);
+	virtual QStringList GetTemplates() const;
+	virtual QStringList GetTextListTmpl(const QString &Setting, const QString& Template) const;
 
 	virtual SB_STATUS InsertText(const QString& Setting, const QString& Value);
 	virtual SB_STATUS AppendText(const QString& Setting, const QString& Value);
 
 	virtual SB_STATUS DelValue(const QString& Setting, const QString& Value);
 
+	virtual QList<QPair<QString, QString>> GetIniSection(qint32* pStatus = NULL, bool withTemplates = false) const;
+
 	virtual SB_STATUS RenameSection(const QString& NewName, bool deleteOld = true);
 	virtual SB_STATUS RemoveSection();
+
+	CSbieAPI*			GetAPI() { return m_pAPI; }
 
 protected: