+++ SandboxiePlus/SandMan/main.cpp
@@ -49,8 +49,6 @@ int main(int argc, char *argv[])
 	CSandMan* pWnd = new CSandMan();
 	QObject::connect(&app, SIGNAL(messageReceived(const QString&)), pWnd, SLOT(OnMessage(const QString&)));
 
-	pWnd->show();
-
 	int ret =  app.exec();
 
 	delete pWnd;